move and rename to generator
